+++ src/utils/npm.ts
@@ -1,0 +1,10 @@
+import { execSync } from 'child_process';
+
+/**
+ * Get the current npm version
+ * @return {String} The npm version
+ */
+export function getNpmVersion(): string {
+  const version = execSync('npm --version');
+  return version.toString();
+}