+++ test/utils/print.test.ts
@@ -6,11 +6,11 @@ import V7_SECURITY_REPORT_TABLE_DATA from '../__mocks__/v7-security-report-table
 
 describe('Print utils', () => {
   it('v6 security report table visual', () => {
-    printSecurityReport(V6_SECURITY_REPORT_TABLE_DATA);
+    printSecurityReport(V6_SECURITY_REPORT_TABLE_DATA, []);
   });
 
   it('v7 security report table visual', () => {
-    printSecurityReport(V7_SECURITY_REPORT_TABLE_DATA);
+    printSecurityReport(V7_SECURITY_REPORT_TABLE_DATA, []);
   });
 
   it('exception table visual', () => {